requirements.txt
pandas>=2.2.2
requests>=2.32.3
numpy>=1.26.0
yfinance>=0.2.44
python-dateutil>=2.8.2
plotly>=5.18.0
kaleido>=0.2.1
python-dotenv>=1.0.0
